a little reorganization (to ease testing)
* index.js becoming more generic as a standard entrypoint.
* index.js requires supportservices.js which exports the necessary view
* separating the underscore template into a separate load step
