+++ package.json
@@ -48,6 +48,7 @@
     "style-loader": "^0.12.0",
     "svg-loader": "0.0.2",
     "underscore": "1.4.4 - 1.8.3",
+    "underscore-template-loader": "^0.7.3",
     "url-loader": "^0.5.7",
     "webpack": "^1.8.5"
   }
