+++ static/templates/services-template.html
@@ -1,0 +1,44 @@
+<h3>Oralington County Support Services</h3>
+
+<div>
+    <div class="pull-left">
+        <div class="services-list">
+            <% var category = ""; %>
+            <% for (var i=0; i < services.length; i++) { %>
+                <% var service = services[i]; %>
+                <% if (category !== service.category.name) { %>
+                    <h4><%= service.category.name %></h4>
+                    <% category = service.category.name; %>
+                <% } %>
+                <div>
+                    <a href="#" data-service-id="<%=service.id%>" class="support-service">
+                        <span class="glyphicon glyphicon-ok" aria-hidden="true"></span> <%= service.title %>
+                    </a>
+                </div>
+                <div><%= service.phone %></div>
+                <br />
+            <% } %>
+        </div>
+        <div>
+            <div class="pull-left">Services Explored: </div>    
+            <ul class="support-service-progress">
+                <% for (var i=0; i < services.length; i++) { %>
+                    <li></li>
+                <% } %>
+            </ul>
+        </div>
+    </div>
+
+    <div class="service-description-list pull-left well" style="display: none">
+        <button type="button" class="close" aria-hidden="true">&times;</button>
+
+        <% for (var i=0; i < services.length; i++) { %>
+            <% var service = services[i]; %>
+            <div class="description" data-service-description="<%=service.id%>" style="display: none">
+                <h5><%= service.title %></h5>
+                <div><%= service.description %></div>
+            </div>
+        <% } %>
+    </div>
+<div>
+<div class="clearfix"></div>
